+++ eqcorrscan/tests/correlate_test.py
@@ -5,7 +5,7 @@ from __future__ import unicode_literals
 
 import copy
 import itertools
-import warnings
+import logging
 from collections import defaultdict
 from functools import wraps
 from os.path import join
@@ -18,10 +18,15 @@ from obspy import Trace, Stream, read
 import eqcorrscan.utils.correlate as corr
 from eqcorrscan.utils.correlate import register_array_xcorr
 from eqcorrscan.utils.timer import time_func
+from eqcorrscan.helpers.mock_logger import MockLoggingHandler
 
 # set seed state for consistent arrays
 random = np.random.RandomState(7)
 
+log = logging.getLogger(corr.__name__)
+_log_handler = MockLoggingHandler(level='DEBUG')
+log.addHandler(_log_handler)
+log_messages = _log_handler.messages
 
 # -------------------------- helper functions
 
@@ -189,10 +194,10 @@ def array_ccs_low_amp(array_template, array_stream, pads):
     for name in list(corr.XCORR_FUNCS_ORIGINAL.keys()):
         func = corr.get_array_xcorr(name)
         print("Running {0} with low-variance".format(name))
-        with warnings.catch_warnings(record=True) as w:
-            cc, _ = time_func(
-                func, name, array_template, array_stream * 10e-8, pads)
-            out[name] = (cc, w)
+        _log_handler.reset()
+        cc, _ = time_func(
+            func, name, array_template, array_stream * 10e-8, pads)
+        out[name] = (cc, copy.deepcopy(log_messages['warning']))
     return out
 
 # stream fixtures
@@ -270,21 +275,22 @@ def gappy_stream_cc_output_dict(
     out = {}
     for name, func in stream_funcs.items():
         for cores in [1, cpu_count()]:
-            # Check for same result both single and multi-threaded
             print("Running {0} with {1} cores".format(name, cores))
-            with warnings.catch_warnings(record=True) as w:
-                cc_out = time_func(func, name, multichannel_templates,
-                                   gappy_multichannel_stream, cores=cores)
-                out["{0}.{1}".format(name, cores)] = (cc_out, w)
+            _log_handler.reset()
+            cc_out = time_func(func, name, multichannel_templates,
+                               gappy_multichannel_stream, cores=cores)
+            out["{0}.{1}".format(name, cores)] = (
+                cc_out, copy.deepcopy(log_messages['warning']))
             if "fftw" in name and cores > 1:
                 print("Running outer core parallel")
+                _log_handler.reset()
                 # Make sure that using both parallel methods gives the same
                 # result
-                with warnings.catch_warnings(record=True) as w:
-                    cc_out = time_func(
-                        func, name, multichannel_templates,
-                        gappy_multichannel_stream, cores=1, cores_outer=cores)
-                out["{0}.{1}_outer".format(name, cores)] = (cc_out, w)
+                cc_out = time_func(
+                    func, name, multichannel_templates,
+                    gappy_multichannel_stream, cores=1, cores_outer=cores)
+                out["{0}.{1}_outer".format(name, cores)] = (
+                    cc_out, copy.deepcopy(log_messages['warning']))
     return out
 
 
@@ -303,20 +309,22 @@ def gappy_real_cc_output_dict(
     out = {}
     for name, func in stream_funcs.items():
         for cores in [1, cpu_count()]:
+            _log_handler.reset()
             print("Running {0} with {1} cores".format(name, cores))
-            with warnings.catch_warnings(record=True) as w:
-                cc_out = time_func(func, name, gappy_real_data_template,
-                                   gappy_real_data, cores=cores)
-                out["{0}.{1}".format(name, cores)] = (cc_out, w)
+            cc_out = time_func(func, name, gappy_real_data_template,
+                               gappy_real_data, cores=cores)
+            out["{0}.{1}".format(name, cores)] = (
+                cc_out, copy.deepcopy(log_messages['warning']))
             if "fftw" in name and cores > 1:
+                _log_handler.reset()
                 print("Running outer core parallel")
                 # Make sure that using both parallel methods gives the same
                 # result
-                with warnings.catch_warnings(record=True) as w:
-                    cc_out = time_func(
-                        func, name, gappy_real_data_template,
-                        gappy_real_data, cores=1, cores_outer=cores)
-                out["{0}.{1}_outer".format(name, cores)] = (cc_out, w)
+                cc_out = time_func(
+                    func, name, gappy_real_data_template,
+                    gappy_real_data, cores=1, cores_outer=cores)
+                out["{0}.{1}_outer".format(name, cores)] = (
+                    cc_out, copy.deepcopy(log_messages['warning']))
     return out
 
 
@@ -358,8 +366,7 @@ class TestArrayCorrelateFunctions:
             assert np.median(cc) != 0.0
             if name == 'fftw':
                 assert len(warning) == 1
-                assert issubclass(warning[-1].category, UserWarning)
-                assert "Low variance found" in str(warning[-1].message)
+                assert "Low variance found" in warning[-1]
 
 
 @pytest.mark.serial
@@ -393,8 +400,7 @@ class TestStreamCorrelateFunctions:
             if cc_name[0:4] in ['fftw_stream_xcorr', 'fftw_multithread',
                                 'fftw_concurrent']:
                 assert len(warning) == 1
-                assert issubclass(warning[-1].category, UserWarning)
-                assert "are there zeros" in str(warning[-1].message)
+                assert "are there zeros" in warning[-1]
         # loop over correlations and compare each with the first in the list
         # this will ensure all cc are "close enough"
         for cc_name, cc in zip(cc_names[2:], cc_list[2:]):
@@ -437,6 +443,7 @@ class TestStreamCorrelateFunctions:
                 np.save("cc_1.npy", cc_1)
                 assert np.allclose(cc_1, cc, atol=self.atol * 100)
 
+
 class TestXcorrContextManager:
     # fake_cache = copy.deepcopy(corr.XCOR_FUNCS)
 
@@ -474,9 +481,13 @@ class TestXcorrContextManager:
 
     def test_str_accepted(self):
         """ ensure a str of the xcorr function can be passed as well """
+        old_default = corr.get_array_xcorr()
+        old_default_stream = corr.get_stream_xcorr()
         with corr.set_xcorr('numpy'):
             func = corr.get_array_xcorr()
             assert func is corr.numpy_normxcorr
+        assert corr.get_array_xcorr() == old_default
+        assert corr.get_stream_xcorr() == old_default_stream
 
 
 class TestGenericStreamXcorr: