+++ eqcorrscan/utils/archive_read.py
@@ -19,13 +19,16 @@ from __future__ import print_function
 from __future__ import unicode_literals
 
 import os
-import warnings
+import logging
 import glob
 
 from obspy import read, UTCDateTime, Stream
 from obspy.clients.fdsn.header import FDSNException
 from obspy.clients.seishub import Client as SeishubClient
 from obspy.clients.fdsn import Client as FDSNClient
+
+
+Logger = logging.getLogger(__name__)
 
 
 def read_data(archive, arc_type, day, stachans, length=86400):
@@ -83,9 +86,12 @@ def read_data(archive, arc_type, day, stachans, length=86400):
 
     .. rubric:: Example, local day-volumes
 
+    >>> # Get the path to the test data
+    >>> import eqcorrscan
+    >>> TEST_PATH = os.path.dirname(eqcorrscan.__file__) + '/tests/test_data'
     >>> t1 = UTCDateTime(2012, 3, 26)
     >>> stachans = [('WHYM', 'SHZ'), ('EORO', 'SHZ')]
-    >>> st = read_data('eqcorrscan/tests/test_data/day_vols', 'day_vols',
+    >>> st = read_data(TEST_PATH + '/day_vols', 'day_vols',
     ...                t1, stachans)
     >>> print(st)
     2 Trace(s) in Stream:
@@ -108,7 +114,7 @@ def read_data(archive, arc_type, day, stachans, length=86400):
         if station_map not in available_stations_map:
             msg = ' '.join([station[0], station_map[1], 'is not available for',
                             day.strftime('%Y/%m/%d')])
-            warnings.warn(msg)
+            Logger.warning(msg)
             continue
         if arc_type.lower() == 'seishub':
             client = SeishubClient(archive)
@@ -124,8 +130,8 @@ def read_data(archive, arc_type, day, stachans, length=86400):
                     channel=station_map[1], starttime=UTCDateTime(day),
                     endtime=UTCDateTime(day) + length)
             except FDSNException:
-                warnings.warn('No data on server despite station being ' +
-                              'available...')
+                Logger.warning('No data on server despite station being ' +
+                               'available...')
                 continue
         elif arc_type.lower() == 'day_vols':
             wavfiles = _get_station_file(os.path.join(
@@ -137,7 +143,7 @@ def read_data(archive, arc_type, day, stachans, length=86400):
     return st
 
 
-def _get_station_file(path_name, station, channel, debug=0):
+def _get_station_file(path_name, station, channel):
     """
     Helper function to find the correct file.
 
@@ -150,13 +156,13 @@ def _get_station_file(path_name, station, channel, debug=0):
     """
     wavfiles = glob.glob(path_name + os.sep + '*')
 
-    out_files = [_check_data(wavfile, station, channel, debug=debug)
+    out_files = [_check_data(wavfile, station, channel)
                  for wavfile in wavfiles]
     out_files = list(set(out_files))
     return out_files
 
 
-def _check_data(wavfile, station, channel, debug=0):
+def _check_data(wavfile, station, channel):
     """
     Inner loop for parallel checks.
 
@@ -166,11 +172,8 @@ def _check_data(wavfile, station, channel, debug=0):
     :param station: Channel name to check for
     :type channel: str
     :param channel: Channel name to check for
-    :type debug: int
-    :param debug: Debug level, if > 1, will output what it it working on.
     """
-    if debug > 1:
-        print('Checking ' + wavfile)
+    Logger.debug('Checking ' + wavfile)
     st = read(wavfile, headonly=True)
     for tr in st:
         if tr.stats.station == station and tr.stats.channel == channel: