+++ CHANGES.md
@@ -1,5 +1,21 @@
 ## Current
+* Changed all prints to calls to logging, as a result, debug is no longer
+  an argument for function calls.
+
+## 0.3.3
+* Make test-script more stable.
+* Fix bug where `set_xcorr` as context manager did not correctly reset
+  stream_xcorr methods.
 * Correct test-script (`test_eqcorrscan.py`) to find paths properly.
+* BUG-FIX in `Party.decluster` when detections made at exactly the same time
+  the first, rather than the highest of these was taken.
+* Catch one-sample difference in day properly in pre-processing.dayproc
+* Shortproc now clips and pads to the correct length asserted by starttime and
+  endtime.
+* Bug-fix: Match-filter collection objects (Tribe, Party, Family) implemented
+  addition (`__add__`) to alter the main object. Now the main object is left
+  unchanged.
+* `Family.catalog` is now an immutable property.
 
 ## 0.3.2
 * Implement reading Party objects from multiple files, including wildcard
@@ -69,6 +85,7 @@
   path on installation of EQcorrscan) that will download all the relevant 
   data and run the tests on the installed package - no need to clone 
   EQcorrscan to run tests!
+
 
 ## 0.3.1
 * Cleaned imports in utils modules